controller refactored
